requirements.txt
gymnasium~=0.29.1
pytest~=8.2.0
numpy~=1.24.3
tensorflow~=2.13.0
docopt~=0.6.2
tqdm~=4.66.2
matplotlib~=3.7.4
pandas~=2.0.3
seaborn~=0.13.1
setuptools~=49.2.1
torch~=2.2.2
typing~=3.7.4.3
gym~=0.26.2
scipy~=1.10.1
tensorboardX~=2.6.2.2
numba~=0.58.1